+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - IssuingStockRule resource
 - rules and debtorWorkspaceId parameters to CreditNote resource
 - reason attribute to PixRequest resource
+- PixFraud resource
 
 ## [0.18.0] - 2026-06-22
 ### Changed